requirements.txt
certifi==2018.10.15
chardet==3.0.4
falcon==1.4.1
gunicorn==19.9.0
idna==2.7
Jinja2==2.10
MarkupSafe==1.0
python-mimeparse==1.6.0
PyYAML==3.13
requests==2.19.1
six==1.11.0
urllib3==1.23
